npm uninstall nodemon watch npm-run-all
We really only needed to watch the server files when we perfected them, now it's not so necessary. Devserver is run just like server now, with `node -r esm`, and this allows us to use imports in that file, and also we can put "type: modules" back in package.json and make VITE happy again. (VITE was warning that using cjs will be deprecated in vite 6)