make map more properly sized, hex translation fix
for hexagon maps, the extra hexagons being added, for larger maps, was such that a whole bunch of dead space was on the left of the map. The translation was too far to the right, but I dont geometrically understand why, yet. The ration is exactly 2! (by the naked eye anyway)