R3F enable on demand frames, MAJOR PERF, disables animations
We don't even have animations yet this is perfect to reduce the battery/cpu load of the game!

https://r3f.docs.pmnd.rs/advanced/scaling-performance#on-demand-rendering
"""
three.js apps usually run in a game-loop that executes 60 times a second, React Three Fiber is no different. This is perfectly fine when your scene has constantly moving parts in it. This is what generally drains batteries the most and makes fans spin up.

But if the moving parts in your scene are allowed to come to rest, then it would be wasteful to keep rendering. In such cases you can opt into on-demand rendering, which will only render when necessary. This saves battery and keeps noisy fans in check.

Open the sandbox below in a full screen and look into dev tools, you will see that it is completely idle when nothing is going on. It renders only when you move the model.
"""